requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
attrs==23.2.0
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
fastapi==0.110.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.6
isodate==0.6.1
lxml==5.1.0
mac-vlc==2020.12.3
macholib==1.16.3
netifaces==0.11.0
numpy==1.26.4
onvif_zeep==0.2.12
opencv-python-headless==4.9.0.80
opuslib==3.0.1
packaging==23.2
psutil==5.9.8
pyaml==23.12.0
pydantic==2.6.4
pydantic_core==2.16.3
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyOgg==0.6.14a1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
scapy==2.5.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.0
uvicorn==0.29.0
valkka-onvif==1.6.7
websockets==12.0
WSDiscovery==2.0.0
zeep==4.0.0
